--- a/src/ZString/Utf16/Utf16ValueStringBuilder.AppendFormat.tt
+++ b/src/ZString/Utf16/Utf16ValueStringBuilder.AppendFormat.tt
@@ -35,7 +35,7 @@ namespace Cysharp.Text
                     if (i != format.Length && format[i + 1] == '{')
                     {
                         var size = i - copyFrom;
-                        Append(format.AsSpan(copyFrom, size));
+                        Append(format, copyFrom, size);
                         i = i + 1; // skip escaped '{'
                         copyFrom = i;
                         continue;
@@ -43,7 +43,7 @@ namespace Cysharp.Text
                     else
                     {
                         var size = i - copyFrom;
-                        Append(format.AsSpan(copyFrom, size));
+                        Append(format, copyFrom, size);
                     }
 
                     // try to find range
@@ -67,7 +67,7 @@ namespace Cysharp.Text
                     if (i + 1 < format.Length && format[i + 1] == '}')
                     {
                         var size = i - copyFrom;
-                        Append(format.AsSpan(copyFrom, size));
+                        Append(format, copyFrom, size);
                         i = i + 1; // skip escaped '}'
                         copyFrom = i;
                         continue;
@@ -84,7 +84,7 @@ namespace Cysharp.Text
                 var copyLength = format.Length - copyFrom;
                 if (copyLength > 0)
                 {
-                    Append(format.AsSpan(copyFrom, copyLength));
+                    Append(format, copyFrom, copyLength);
                 }
             }
         }